more enhancement in friend api

+++ friend/views.py
@@ -20,41 +20,51 @@ class SendFriendRequestView(APIView):
     def post(self, request):
         serializer = FriendShipSerializer(data=request.data)
         if serializer.is_valid():
-            receiver_name = serializer.validated_data.get('receiver').name
+            receiver = serializer.validated_data.get('receiver')
+            receiver_name = receiver.name  # Assuming `name` is a field in the receiver model
+            # receiver_name = serializer.validated_data.get('receiver').name
+
             # Set the sender as the logged-in user
-            friend_request = serializer.save(sender=request.user)
-            return Response({'status': 'success', 'message': f"Friend request has been sent successfully to {receiver_name}", 'friend_request_id': friend_request.id}, status=status.HTTP_201_CREATED)
+            friend_request = serializer.save(sender=request.user, receiver=receiver)
+            return Response({
+                'status': 'success',
+                'message': f"Friend request has been sent successfully to {receiver_name}",
+                'friend_request_id': friend_request.id,
+                'receiver': receiver.id
+            }, status=status.HTTP_201_CREATED)
         else:
             errors = serializer.errors
             return Response({'status': 'error', 'message': 'Failed to send friend request', 'errors': errors}, status=status.HTTP_400_BAD_REQUEST)
-
-
 
 class AcceptOrRejectFriendRequestAPIView(APIView):
     permission_classes = [IsAuthenticated]
 
     def post(self, request):
-        friend_request_id = request.data.get('friend_request_id')
+        serializer = FriendShipSerializer(data=request.data)
         
-        try:
-            friend_request = FriendShip.objects.get(id=friend_request_id)
-        except FriendShip.DoesNotExist:
-            return Response({'message': 'Friend request does not exist.'}, status=status.HTTP_404_NOT_FOUND)
-        
-        # Perform any additional checks or validations here
-        
-        action = request.data.get('action')
-        
-        if action == 'accept':
-            friend_request.accept()
-            return Response({'message': 'Friend request accepted.'}, status=status.HTTP_200_OK)
-        elif action == 'reject':
-            friend_request.reject()
-            return Response({'message': 'Friend request rejected.'}, status=status.HTTP_200_OK)
+        if serializer.is_valid():
+            receiver = serializer.validated_data.get('receiver')
+            
+            try:
+                friend_request = FriendShip.objects.get(receiver=receiver, sender=request.user)
+            except FriendShip.DoesNotExist:
+                return Response({'message': 'Friend request does not exist.'}, status=status.HTTP_404_NOT_FOUND)
+            
+            # Perform any additional checks or validations here
+            
+            # action = serializer.validated_data.get('action')
+            action = request.data.get('action')
+            
+            if action == 'accept':
+                friend_request.accept()
+                return Response({'message': 'Friend request accepted.'}, status=status.HTTP_200_OK)
+            elif action == 'reject':
+                friend_request.reject()
+                return Response({'message': 'Friend request rejected.'}, status=status.HTTP_200_OK)
+            else:
+                return Response({'message': 'Invalid action.'}, status=status.HTTP_400_BAD_REQUEST)
         else:
-            return Response({'message': 'Invalid action.'}, status=status.HTTP_400_BAD_REQUEST)
-
-
+            return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
 
 
@@ -84,32 +94,40 @@ class ViewAllFriendRequestAPIView(APIView):
             'friend_requests': serialized_data
         }, status=status.HTTP_200_OK)
 
+
+
 class UnfriendAPIView(APIView):
     permission_classes = [IsAuthenticated]
 
     def post(self, request):
-        user_id = request.data.get('user_id')
+        receiver = request.data.get('receiver')
         
         try:
-            friendShip = FriendShip.objects.filter(sender=request.user, receiver__id=user_id).first() \
-                         or FriendShip.objects.filter(receiver=request.user, sender__id=user_id).first()
+            friendShip = FriendShip.objects.filter(sender=request.user, receiver=receiver).first() \
+                         or FriendShip.objects.filter(receiver=request.user, sender=receiver).first()
         except FriendShip.DoesNotExist:
-            return Response({'message': 'FriendShip does not exist.'}, status=status.HTTP_404_NOT_FOUND)
+            return Response({'message': 'Friendship does not exist.'}, status=status.HTTP_404_NOT_FOUND)
         
         friendShip.delete()
         
         return Response({'message': 'Unfriended successfully.'}, status=status.HTTP_200_OK)
 
-
-
-
 class CancelFriendRequestAPIView(APIView):
     permission_classes = [IsAuthenticated]
 
-    def delete(self, request, request_id):
-        try:
-            friend_request = FriendShip.objects.get(id=request_id, sender=request.user)
-            friend_request.delete()
-            return Response({'message': 'Friend request canceled successfully.'}, status=status.HTTP_204_NO_CONTENT)
-        except FriendShip.DoesNotExist:
-            return Response({'message': 'Friend request not found.'}, status=status.HTTP_404_NOT_FOUND)
+    def delete(self, request):
+        serializer = FriendShipSerializer(data=request.data)
+        
+        if serializer.is_valid():
+            receiver = serializer.validated_data.get('receiver')
+
+            friend_request = FriendShip.objects.filter(receiver=receiver.id, sender=request.user).first()
+
+            if friend_request:
+                friend_request.delete()
+                return Response({'message': 'Friend request canceled successfully.'}, status=status.HTTP_204_NO_CONTENT)
+            else:
+                return Response({'message': 'Friend request not found.'}, status=status.HTTP_404_NOT_FOUND)
+        else:
+            return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+
